requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
linkedin_api==2.1.1
openai==1.28.1
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
pymongo==4.7.3